Some simple Enzyme and Jest Mock tests. For some reason even when using enzyme's mount() it complains about GridWindow having setState used, which a little annoying.
Apparently I cannot use setState in these tests which happens during construction. This is a necessity because setState is what triggers re-rendering of the various elements when the tags, genres, and ratings are successfully retrieved. If I'm not allowed to do that then I'd have to write some manual re-draw functions to bypass setState when I'd really rather just use setState.

Documentation and Googling is spotty on this issue, it's hard to find anyone talking about this issue and in only one location did I find someone explaining that one had to use enzyme's mount instead of shallow - and it didn't solve my problem:
https://github.com/facebook/jest/issues/3235