Client: Edit Movie tool is working though very buggy. Fixed some state mutatation issues around checkedTags by using cloneDeep. There are articles talking about solutions such as helpers described in this: https://medium.com/pro-react/a-brief-talk-about-immutability-and-react-s-helpers-70919ab8ae7c Added a "queryOK" utility function to replace some repeated code - could use it elsewhere but don't have the time. Ran into a weird GraphQL/Apollo bug where the cache was getting confused. I apparently solved this by making sure to always grab the id in every query for everything I request - something I should be sure to do from now on. Note to self: Use id all the time whenever possible and derive UI from them (avoid strings as an index in memory). See genre and rating issues in MovieEditor > Basics > Rating Element.
Bugs: Switching between routes is buggy as hell.
If you navigate from movie/edit to movie/create the UI won't update. Navigating the other way is also bad.
Sometimes you need to hard refresh the page to populate the edits.
Really, routing is a mess with state and queries and cache - it never clears when it should. It refreshes randomly. Sometimes it fails to even populate correctly.
This project is starting to become a mess, to be honest. I mean, it's not for production or anything but I definitely need to learn some lessons from this whole thing.