Client: Replacing UserWrapper with UserProtector. Essentially swapping out a wrapped component with a Higher Order Component (HOC) and named it withUser(), adding a slipped-in props forwarder.
Trying to figure out how to make client login/logout play correctly with Apollo. Turns out that UserWrapper was actually a totally known pattern that I thought was terrible.

Going to need to continue research. Perhaps I need to deal with updating the Apollo cache? I need to figure out a way to make the HeaderMenu realize that the user is no longer logged in - and I'm hoping I can find a way that doesn't involve refreshing the page.