Client: Added cache modification on the event of adding and updating movies. I got stuck for too long with the add functions because I didn't realize that the name of the second object needed to be "data" - that's what I get for using object name shorthand notation!
I consider this project feature complete. Sure, I could've added pagination and a lotmore, but that's polish - this is just a practice project that I took on to make my first crack at this a messy, ugly mess - and I've learned a lot about ReactJS, Routes, async, npm/yarn, REST, GraphQL, Prisma, and Apollo Client. This is the minimum viable product and I'm ready to move onto getting this thing dockerized and hosted online like a real website.