Client: Using port forwarding in the compose to use port 3005 on host instead of setting the docker container to use port 3005. So now it just goes off port 3000.
Server: CMD yarn start doesn't work, it runs into issues. Perhaps to do with nodemon? Anyway, I've set it to just start the normal way but I'm struggling to get it to talk to the Prisma container. And it's bad timing that I updated Prisma while working and now I cannot restart the container. Will probably have to Nuke it.